bugfix alignment lost
it had been expecting a key `id` to be available, but sometimes it's referenced as the full length `directory_id`. this can be from backend api changes, caching etc. it's mostly just a legacy thing we are trying to get rid of so for now this will work for either case.

code direction is actually for each directory to be an object in the SDK which will likely require changing this again